requirements.txt
openai
sqlalchemy
mysql-connector-python
colorama